moved bank modes from users file
