+++ flite/banks/models.py
@@ -1,3 +1,28 @@
 from django.db import models
+from django.contrib.auth import get_user_model
 
 # Create your models here.
+
+User = get_user_model()
+
+
+class AllBanks(BaseModel):
+
+    name = models.CharField(max_length=100)
+    acronym = models.CharField(max_length=50)
+    bank_code = models.CharField(max_length=50)
+
+    def __str__(self):
+        return self.name
+
+    class Meta:
+        verbose_name_plural = "All Banks"
+
+
+class Bank(models.Model):
+
+    owner = models.ForeignKey(User, on_delete=models.CASCADE)
+    bank = models.ForeignKey(AllBanks, on_delete=models.CASCADE)
+    account_name = models.CharField(max_length=100)
+    account_number = models.CharField(max_length=50)
+    account_type = models.CharField(max_length=50)
