+++ flite/users/models.py
@@ -93,27 +93,6 @@ class Balance(BaseModel):
         verbose_name= "Balance"
         verbose_name_plural = "Balances"
 
-class AllBanks(BaseModel):
-
-    name = models.CharField(max_length=100)
-    acronym = models.CharField(max_length=50)
-    bank_code = models.CharField(max_length=50)
- 
-    def __str__(self):
-        return self.name
-
-    class Meta:
-        verbose_name_plural = "All Banks"
-
-
-class Bank(models.Model):
-
-    owner = models.ForeignKey(User, on_delete=models.CASCADE)
-    bank = models.ForeignKey(AllBanks, on_delete=models.CASCADE)
-    account_name = models.CharField(max_length=100)
-    account_number = models.CharField(max_length=50)
-    account_type = models.CharField(max_length=50)
-    
 class Transaction(BaseModel):
     owner = models.ForeignKey(User, on_delete=models.CASCADE, related_name='transaction')
     reference = models.CharField(max_length=200)
@@ -136,7 +115,6 @@ class P2PTransfer(Transaction):
 
     class Meta:
         verbose_name_plural = "P2P Transfers"
-
 
 
 class Card(models.Model):
